Moved PotionColor out of the Brew class.
